created the endpoint for findCategory and updateById for movie

+++ src/main/java/com/dravinck/movies/controller/MovieController.java
@@ -10,6 +10,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+import java.util.Optional;
 
 @RestController
 @RequestMapping("movies/movie")
@@ -33,17 +34,37 @@ public class MovieController {
     }
 
     @GetMapping("{id}")
-    public ResponseEntity<MovieResponse> findById(@PathVariable Long id){
+    public ResponseEntity<MovieResponse> updateById(@PathVariable Long id){
         return movieService.findById(id)
                 .map(movie -> ResponseEntity.ok(MovieMapper.toMovieResponse(movie)))
                 .orElse(ResponseEntity.notFound().build());
     }
 
+    @PutMapping("{id}")
+    public ResponseEntity<MovieResponse> updateById(@PathVariable Long id, @RequestBody MovieRequest request){
+        return movieService.update(id, MovieMapper.toMovie(request))
+                .map(movie -> ResponseEntity.ok(MovieMapper.toMovieResponse(movie)))
+                .orElse(ResponseEntity.notFound().build());
+    }
+
+    @GetMapping("/search")
+    public ResponseEntity<List<MovieResponse>> findByCategory(@RequestParam Long category){
+        return ResponseEntity.ok(movieService.findByCategory(category)
+                .stream()
+                .map(MovieMapper::toMovieResponse)
+                .toList());
+    }
+
     @DeleteMapping("{id}")
     public ResponseEntity<Void> delete(@PathVariable Long id){
-        movieService.delete(id);
-        return ResponseEntity.noContent().build();
+        Optional<Movie> optMovie = movieService.findById(id);
+        if (optMovie.isPresent()){
+            movieService.delete(id);
+            return ResponseEntity.noContent().build();
+        }
+        return ResponseEntity.notFound().build();
     }
+
 
 
 }
